requirements.txt
pyspark
kafka-python
pandas
boto3
faker
